+++ src/helpful-queries/all-builds-with-item.sql
@@ -5,3 +5,12 @@ FROM
   JOIN BuildItems ON Build.id = BuildItems.buildId
 WHERE
   BuildItems.itemId = 'ITEM_ID';
+
+#-------------------------------------
+SELECT
+  COUNT(Build.id) AS buildCount
+FROM
+  Build
+  JOIN BuildItems ON Build.id = BuildItems.buildId
+WHERE
+  BuildItems.itemId = 'ITEM_ID';